+++ climpred/__init__.py
@@ -14,6 +14,7 @@ from . import (
     tutorial,
 )
 from .classes import HindcastEnsemble, PerfectModelEnsemble
+from .preprocessing import mpi, shared
 from .versioning.print_versions import show_versions
 
 try: